# md4: Optimize compress to improve hash performance by about 12.5%