@newpavlov 2023-11-18
Thank you! On my PC performance improvement is much smaller (~2.5%), but still measurable.